renames session manager classes to message history
