+++ redisvl/extensions/message_history/__init__.py
@@ -1,0 +1,7 @@
+from redisvl.extensions.message_history.base_history import BaseMessageHistory
+from redisvl.extensions.message_history.message_history import MessageHistory
+from redisvl.extensions.message_history.semantic_message_history import (
+    SemanticMessageHistory,
+)
+
+__all__ = ["BaseMessageHistory", "MessageHistory", "SemanticMessageHistory"]
